wlan: trim the meta-information from logging packets
The 4 byte meta-information attached by firmware to
the logging packets is used to classify them into different
log types so that they can be displayed correctly. But the
logging applications do not expect the meta-information as
part of the payload, so we have to trim it out before passing
it to the user-space.

CRs-Fixed: 823982

+++ CORE/SVC/src/logging/wlan_logging_sock_svc.c
@@ -836,7 +836,7 @@ int wlan_logging_sock_deinit_svc(void)
        return 0;
 }
 
-int wlan_queue_logpkt_for_app(vos_pkt_t *pPacket, uint8 pkt_type)
+int wlan_queue_logpkt_for_app(vos_pkt_t *pPacket, uint32 pkt_type)
 {
 	unsigned long flags;
 	vos_pkt_t *next_pkt;
